dodanie niektórych kategorii
na razie tylko do wyglądu
