+++ src/pages/statistics/index.js
@@ -1,78 +1,21 @@
 import { Box } from "@mui/system";
-import ListSubheader from "@mui/material/ListSubheader";
-import List from "@mui/material/List";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import Collapse from "@mui/material/Collapse";
-import InboxIcon from "@mui/icons-material/MoveToInbox";
-import DraftsIcon from "@mui/icons-material/Drafts";
-import SendIcon from "@mui/icons-material/Send";
-import ExpandLess from "@mui/icons-material/ExpandLess";
-import ExpandMore from "@mui/icons-material/ExpandMore";
-import StarBorder from "@mui/icons-material/StarBorder";
-import { useState } from "react";
-import { Button } from "@mui/material";
+
+
+import { Button,Checkbox } from "@mui/material";
 import ChartWithMenu from "../../components/statistics/professions/ChartWithMenu";
 import { Professions } from "../../components/statistics/professions/Professions";
+import MenuBar from "../../components/statistics/MenuBar";
 
 export function Statistics(props) {
-  const [open, setOpen] = useState(true);
-  const handleClick = () => {
-    setOpen(!open);
-  };
+
   return (
-<>
-<Box sx={{width:'100%',height:'auto',display:'flex'}}>
-      <Box sx={{ width: 300, height: 300,border:'2px solid black', float:'left'}}>
-        <List
-          sx={{ width: "100%", maxWidth: 360, bgcolor: "background.paper" }}
-          component="nav"
-          aria-labelledby="nested-list-subheader"
-          subheader={
-            <ListSubheader component="div" id="nested-list-subheader">
-              Wybór kryteriów
-            </ListSubheader>
-          }
-        >
-          <ListItemButton>
-            <ListItemIcon>
-              <SendIcon />
-            </ListItemIcon>
-            <ListItemText primary="Sent mail" />
-          </ListItemButton>
-          <ListItemButton>
-            <ListItemIcon>
-              <DraftsIcon />
-            </ListItemIcon>
-            <ListItemText primary="Drafts" />
-          </ListItemButton>
-          <ListItemButton onClick={handleClick}>
-            <ListItemIcon>
-              <InboxIcon />
-            </ListItemIcon>
-            <ListItemText primary="Inbox" />
-            {open ? <ExpandLess /> : <ExpandMore />}
-          </ListItemButton>
-          <Collapse in={open} timeout="auto" unmountOnExit>
-            <List component="div" disablePadding>
-              <ListItemButton sx={{ pl: 4 }}>
-                <ListItemIcon>
-                  <StarBorder />
-                </ListItemIcon>
-                <ListItemText primary="Starred" />
-              </ListItemButton>
-            </List>
-          </Collapse>
-          <Button>Zatwierdź</Button>
-        </List>
+    <>
+      <Box sx={{ width: "100%", height: "auto", display: "flex" }}>
+       <MenuBar/>
+        <Box sx={{ width: "auto", float: "left", flexGrow: 1 }}>
+          <Professions />
+        </Box>
       </Box>
-      <Box sx={{ width:'auto',float:'left',flexGrow:1 }}>
-          <Professions/>
-      </Box>
-      </Box>
-
-
     </>
   );
 }
